refactor shexec parser.go into new package shellenv.  separate out bash specific parsing from generic functions
